services/metrics-api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
clickhouse-driver==0.2.7
pika==1.3.2
python-dotenv==1.0.1
